requirements.txt
fastapi==0.115.7
fastapi-mail==1.4.2
uvicorn==0.34.0
celery==5.4.0
eventlet==0.38.2
redis==5.2.1
requests==2.32.3
SQLAlchemy==2.0.37
pydantic==2.10.5
python-dotenv==1.0.1
psycopg2-binary==2.9.10
